Update 14 for strategy game
Added support for switching modes for a unit. the "a" key switches a unit to attack, displaying its attack range, while the "m" key switches a unit to move range, displaying its move range. A unit should be able to toggle between modes at any point when it is selected.