Update 15 for strategy game
Added a "will_fit" method to restrict a units movement to ensure that its entire sprite (all 4 tiles) must fit in the tile its selected to move to. This fixed the problem where a unit could move farther right and down than it could left and up, since the only check was for its upper right corner fitting in the move range.